better handling for non-HTML pages
instead of throwing an exception have `to_page` return an instance of a
different class (`NonHtmlPage`) that returns `nil` for `title`,
`description` etc.

allows different media types to be treated interchangeably, and
potentially sets the stage for extracting titles and descriptions from
other file formats such as PDF in the future.