improve permanent/temporary redirect handling
when returning the URL in the result of `RFetch.get` we only want this
to take account of _permanent_ redirects. We still follow _temporary_
redirects in order to arrive at the eventual content to return, but
the URL returned should be either the originally requested URL (when
there are no redirects) or the end of a chain of permanent redirects
before encountering any temporary redirects.

i.e. get A, permanent redirect to B, temporary redirect to C returns the
URL of B and the content from C.