include Accept header in requests
prompted by an issue where requesting a file from GitHub would
return JSON rather than HTML content

e.g. for https://github.com/oxidecomputer/oxide-and-friends/blob/master/2021_07_05.md

Narrowed this down to lack of an Accept header. Should prioritise the
HTML version of content so that we’re able to extract the title and
description, but also it feels that in general we should try to make
the request close to what a browser would get back.